2024/09/10 22:04 trial and error a lot of error!
It is kind of was a failure... I was trying to add an indicator almost.. no an entire day, but without kind of a success? I still need to work on it a lot... OP was kind enough to help me out, and I appreciate him because of it, but I kinda feel down! He talked about that projects shouldn't be over complicated, and I agree... so why am I trying to over complicate this project? Honestly idk I want it to be neat and tidy so I can understand and change almost any part of the code at my will. But at the same time I'm doing just like that over complicate the task, I wasn't even able to finish a task today which is alright I don't have to finish everything at once, but it gets me thinking. I'm locked in for now and want to continue, but to be honest if this amount of failure was enough to discourage me than I'm kind of scared if I even ready for what else to come. If even this small of a task of "get an indicator into a ellipse" is enough to give me a hard time, than I feel this will be 10x harder when I get to attacks, EMP, and oh boi enemies... I'm scared, but if I got to this far. Fuck me no one should stop me. (IF I get karma-ed bc of this... pls kill me )