2024/09/12 11:40 UI start and movement upgrade :D
Alright I finished the Indicator finally yesterday! It works a bit weird at the moment, but I can figure it out something later. For now I want to focus on just adding stuff in and making them at least on a working state. If I feel okay about them at least than that's mean I'm done with that feature for the time being! I have also done around 1/3 of the UI, which just means that I made the UI part for the ABIL, ACT, BAG, RUN! they look okay which mean I can jump to the next feature, which is going to be the turns... YAY ... XD

Any how I feel like for now it shouldn't be so bad... I will figure it out when I'm  there for now that's all what I did the past 2 days. I will try to start it around the afternoon or around nightfall so I at least have some idea how to start. Other than that I tried to make a custom font that is coming along somehow, and the BG for the trees are still hunting me atm. *sigh* Anyways keep you update later!