2024/09/14 INITIATIVE is DONE
Heh it was easier than I thought this was the first feature that I actually enjoyed coding it in. Not like the other was worse, but they gave me a tiny bit harder time. But this was basically a cake in a park! And everything worked out just h ow I wanted it. Although I would like to test it later down the line for more RNG checks bc I'm ain't 100% sure it's working perfectly, but for now it is done. Next will be ENG and HP meter. Kinda excited to do it but for now I want to eat and chill, bc I only wake up 4 hours ago. Anyhow stuff in my mind, but the time you show this to someone this section will be irrelevant and a long gone thought. And I hope everything worked out in the end between you too. And you should know what I'm talking about... if not (Andrew Flower)... sorry if was that bad. Anyways I hope you like explaining yourself :D