2024/09/19 16:03 Health bar
Alright I did it... kinda! I still need to work on the numbers when one of the bar gets damaged. For some reason the function duplicates and removes twice as much of points so that's ahead. If you watched on the date you could probably guessed that is been 5 days since I worked on the project around any amount. The simple answer is that I had a rough week. Not going into details too much details bc maybe someone else's reading this. But shortly I got tired some and needed a tiny bit of break. BUT now I'm back and ready to continue and wanting to make this the best I could! For now I go back to chill, I will try to make some changes tomorrow so I'm not completely forgetting this project, but I need some rest if I want to or not!