2024/09/19 22:32 Bar system working YAY
It is done YAY! I did it and it is looking good. Tomorrow hopefully I will start to make a dummy enemy and after that we can start adding the actual challenging stuff. Such as: The Action Button

YAY... okay I'm not that excited bc I now something will break almost immediately, but let it be known that I will try my best to make it as awesome as I can do it!