2024/09/22 21:42 small set back
Yesterday I realized something with how I wanted to structure my code. I created 1 main way to control both enemies and players, however this approach was shit to say the least. The truth is I spent way to much time on editing that one file and trying to get the player movement right that I forgot that I also wanted to use it for enemy behaviour. And because I already implemented the player movement I couldn't add anything else. Put it together that yesterday I had a pretty shitty day (mentally). I deleted some stuff. Nothing much I was able to recover the character.tscn and from there I rewrote it so small portion is shared with both enemy and player, but they have 2 different controls in the end. I also added consistent HP and ENG tracking... it is not a big deal, but for a while I was trying to avoid to use any giant _process functions, to avoid optimisation hell later. But I was overstressed on this one ... it is comically shit to be afraid of something that is a main part of godot! Anyways trying to create a dummy atm and working my way towards action, my first (in my mind) bigger road block. The simple reason is bc I have to start considering a lot more and I have to start adding actual features which needs to interact with an enemy AND I soon have to find an easy way to display text... so in my mind it is going to be the first actual challenge of this project... which not really a good sign if I had trouble with the considerably easier part of the project! But this is how programming it is. I will looking forward, but I might put it onto the back burner for maybe around 1-2 days bc next Friday (2024.09.27) I will be able to showcase Collision my previous project to some Elementary students, which I'm excited for! So I want to clean up a tiny bit more before I jump into it! Honestly that's all... it has been quite the long day so I want to go take rest soon, but I'm in a better headspace right now like I was yesterday or even the entire week, if I can say so myself. I had fun, we had laughs and I feel a bit okayer. Let's keep record how long until I feel shit again. READY? GO