2024/09/23 16:31 Actions are half way programmed
I had to call it for the day early bc I have some stuff to do. I have another game to polish a bit before Friday or Thursday or when I have to go to my school and show my previous project! The truth is that there is a lot that I have no idea atm and a lot of stuff feel like I need to right now and Immediately! Which makes me angry and to be honest I just want throw everything in the air from anger and just walk out of a door. It is shit to think about stuff which half of them not even have a date or any reason to exist. I'm kinda overwhelmed of all the stuff that just happened TODAY. But at the same time I couldn't care any less. All of these are bullshit one way or another! BUT actually back to the update XD.

I created a way to create action options and tie them back to the name of the enemy! From there when I try to interact with it it brings up all the options to that is tied to his name and the next move is when one of those buttons is pressed it plays one of its custom functions! I'm kinda overwhelmed so I need to chill and hopefully I will be back in either Wednesday or Thursday to continue. For Now It Is CHILL!