2024/09/28 19:21 Action cancel
Took some days off, not to rest that's for mostly sure. But I wanted to work on a bit more before this week is over. I love how it looks so far. Next week is going to be fun or boring depending on some stuff. But hopefully next week I will be able to work on it more and will be able to add flexible buttons. Ohuhohu I would be lying If I wasn't a bit excited about flexible buttons. But for now I want to chill and relax. Other than that I have nothing else to report GN!