2024/10/01 10:07 Act is done!
Oh yeah I suppose to publish this yesterday, oh well. Anyways I finished act, it isn't the best I would've liked but it is better than using if's every time! Now the 2 things up to do is Camera movement... and Dialogue... TBH I WANT TO GO BACK TO DO ACT... XD okay for real these 2 are my biggest fear for now, bc Camera movement isn't entirely bad, but the fact that  I have to do it almost alone is the scary stuff, found some assets online that could be useful later, but the ONE FETURE that I wanted to use is bugged and returns an error so I cannot use it! On the other hand The dialogue is scary bc it is the most complicated part of the entire project. Lines of dialogue is basically  half of the game and if I have something that is a pain in the ass to work with than I WILL gonna have a bad time! The immediate idea that just got in my mind is that I could just not use that many dialogues but if I cannot find any good solutions today, than I might consider it more. Next update defiantly is going to be one or the other, or another option we will see! No rent today... you don't have to listen to that XD.