2024/10/02 14:36 Working Camera
Well it was an interesting experience! I was able to make it quite good although I will need feedback on it and more testing later, we will see, but for now it works and that is the main goal! The camera is just an asset I downloaded, it has a bunch of useful tools such as targets, path cams, zoom outs and camera rooms where I can stuck the camera in a "room" and it stays inside those boundaries. How I made that into a mouse following camera was quite tricky, bc I need to disable and hide all ui for better see ability and for the player not to press a button and not see what they did. I also added zoom which is just adds like 0.1 every mouse scroll. It is neat but I would've loved if it would be finished like yesterday, really shows how much struggle I'm having with the simplest features, and it does not help the next is basically an entire BEAST of an feature! Dialogues are terrifying yeah I'm using word I can hardly write down that's how you know I'm serious. To be real found a really interesting asset, and I plan to add it, but the negative side of that which is 2 is:
- It is using its own language which I need to learn if I choose it.
- When I used it on the asset dump project it just didn't work the tester, which is scary bc I couldn't really find on first glens what is the problem and I do not really want to debug an asset if that is an option. I will try it out later, for now wanna take a break cause I got a really rough morning and although this fun to debug, I need to do house work... I know y'all in the future don't give five shit just let past RTS vent for shit! XD  Anyways will update you later what happens next :)