2024/10/06 11:13 Starting with Dialogue
Alright! This is... well... I'm not really good with words to describe how this felt. The truth is that it was not easy, but although that was expected I thought that I will going to have problems with... the other stuff such as multiple dialogue boxes, variable handling and other stuff, and yes I haven't done them yet, but by itself the asset I use on its own giving me trouble! I had problem even to use it through code and just make it so it fits neatly to the other part of the code... But at this point I only managed to get it to work with basic text stuff, show and hide and show it when the act option is on or off and around that's it. It is complicated, but I feel it is not impossible and quite the other hand I feel it could help me big time in the long run IF I learn how to use it properly which is why it is hard. The hard part is just on its way if for me to get it to work was around 2-3 different errors and around 2 hours of just debugging was needed to figure it out than you can probably guess how much I give myself for the other stuff to work out flawlessly...

And yeah I know I'm not sounding too optimistic, well there is nothing wrong I'm just idk a bit empty headed atm! Excited? Yes! Knowing I can do this? Also Yes! Knowing it will take time? Of course Yes! Although I'm not sounding bouncing happy of thought of doing bonus work because of this stupid dialogue not working 100% with me... I'm still hoping for this project to be the best it can be, and for now I'm still having that excitement just knowing that this is a gatekeeper for a bunch of the games content that around 79% players will just skip or don't care or knows it already. So it needs to count for not all but for a good margin to be good, snappy and memorable! I will see in the future and keep you updated, but for now that's all. Hoping to commit next around.. idk... Tuesday? We will see!