2024/10/10 17:53 Dialogue and comments!
Alright, alright you might be a future me who knows everything BUT MAN this stupid Dialogues are killing me, I was kinda able to progress but not too much!. So from last time so you are not escaping the rent! I switched from one dialogue plugin to another! Which is just boils down to I'm having the best of my troubles and just cannot understand most stuff that people doing bc it is not mine!. Pff I really don't like to work with assets they not make my life and time easier but make me backtrack to fix issues that others haven't fixed yet. I would ask for help from OG and I want to in the upcoming days hopefully, but for now I will work on my own. I haven't even touched a great portion of the list of things I have to do. This is still JUST the beginning! pfft And the kinda sad part is that I slowly have to put the project down bc of learning and stuff, I'm still at the beginning of the year so I have time until December / January , but the time is running out and I day by day losing interest which makes me scared that I will not be able to finish THE TEST of the mechanics! Looking back on me around 2 weeks ago I with still high hopes and me right now being just tired and scared that I have to throw in the towel later down the line bc of something is really scary, I don't want to give up yet! Not yet... I'm in a valley right now if you know what I mean! It will be hard to get back to the excitement of the previous month, but I'm still standing strong both physically and emotionally not to give up. I feel if I can stand strong and even if a little over time I can get it to work. There are still a bunch of time before me so I shouldn't worry about it too much just the fact I want to make it sooner than later, you know?

Any ways! I talked WAY TOO Much right now so I was able to fill up last and now's rent. I hope as always it was worth it. tomorrow I will commit another one without rent bc something stopping me to quite out so I have to write this now and finish the commenting and other stuff in a second part. Any how I only hope that the future me can just only laugh at this and smug about it that all of this was worth it and I haven't given up on this project and everything else!... Please keep up me... We know we can do everything if we put our mind together and that count for you too reader ;) Have a great day!