2024/10/15 21:00 Comment Checkpoint
I haven't been great lately, so if you read that one where you I asked to remember when I will feel bad.. than I guess now is the time to cash it in. I'm just demotivated, not even because of the project, but because the stuff around me, I'm tired confused and just not in a good mood in general! I want to work on it today, but just didn't... which is just a shitty mind set at all! I want and will continue because even tho I'm far from over with it, it is fun to work on it, I just need to find motivation and time again and than I will be back on it again with lightning speed! I will try to post like every 2-3 days as usual, but it is hard to get myself to it... idk I just don't feel anything idk it is weird to describe it and I don't even know if I have the vocabulary to do so. XD I took like a week to comment stuff you know there are some actual real shit in the background. I will keep you update later. And I HOPE this will not be the last update I will post... well not really bc I wanna continue and as long as I have that 'want to' I hope thar 'I can' do it! anyways