2024/10/30 23:39 Menu Rework and finally back
Oh hey It's been a while... Not for you that's for sure, for me I defiantly took at least 2 weeks to get back to the project, I'm refreshed I did a my first game jam and got some new refreshing thoughts I was able to rethink how the entire menu system would look like and I was even able to work out the prototype and right now I can say it with a 100% guarantee that the new module is super functional. Gone the days of headache and now it is super see through able I didn't even needed OP's help even though I would've appreciated it, but regardless I did it, I made it compatible with the game and as I stand it is functional, it isn't 100% done there are many bugs and problems I'm curtain, but as where I stand it is ay-o-kay! You will not gonna get any bedtime stories right now even tho I could vent a bunch, I'm just happy that I was able to progress and I don't wanna talk about anything other than this. Anyways next stop is to make the run and the EMP functions work, I do not see any threats atm merely because it isn't much of a hustle like Actions, Abilities or Items it should be easy... Alright that's all for the night and see you in the next post! :D