2024/10/31 19:04 REALLY small update
Alright, I know I haven't done too much this time in the game, but I wanted to check on it before I forget about it again for a good week. I did some hp bar scaling and I managed to work out the basis of the run mechanic, I think it is done for now until I need to add it with victory and Mercy victory. The last 4 before the Mid Game section of the mechanics are:
- Bagpack
- Items
- EMP meter
- Damage Indicator
I'm not scared of EMP meter and Damage Indicator because they seem easy and kinda forward (the only reason I didn't do EMP meter today is because I have no idea how to approach the design, aka I'm lazy af atm XD), But the other 2 seem... like the final boss of this section. I did Dialogue, Camera, Menu, UI but none of them is comparable to Item and its management... There are countless tutorial of it and a bunch of plugins on it and almost all of them requires some sort of data storage for your items. Which I'm not good at... I like to play around with data's and storing smaller numbers but this is the largest scale I have ever done and it is my first time doing stuff like this!. So even if I somehow wrong about it I will going to learn a lot from this last one and I feel like this is going to be a section boss for sure!

The reason why I say this is bc I divided my task to 3-4 sections from early setup to endgame storing and enemy AI. As I go forward the task gets more harder and harder for me. I already worked 2 months at this point and I just got past the earliest phases of he project from here it is all shits but no giggles. And I can only say this much... If this took 2 months... Than make sure to double it MF!