2024/11/28 16:37 Ability and early begining
Alright right now it is almost another week since I left early and started working on the next phase of the project! I've been working slowly but surely on the ability sections and I feel like I'm making steady progress! It's good to meet other people that are interested in my project and I can actually make good progress. I also feel like I'm getting to those people that have the same interest as I do it is weird to find people that are so enthusiastic like I'm and just want to know more encourage me to make it and continue I have no clue what changed in this month but since I took the game jam I feel nothing but joy to sit down and work on the project and although it is slow, I do make progress and I just cannot help the feeling of proudness and hope for getting this done before February! LIKE THAT IS A BIG SAY but I think it is possible, but I'm still far away and also holidays are coming! BUT and that is a big BUT because I have a feeling that soon I will be done with my Technicum project and than I only need to learn like 20 different thesis and I will be done for this year! SOoo... I will slowly have more and more time to work on this and this is where I usually would fail and procrastinate on this for days, weeks or months, and here is the BIG BUT! Things changed and what felt like impossible and unmoveable is now replaced with feeling like encouraged and plausible in correct matter  and I can only kinda hope that this will last just enough for me to finish and before the end of the school year unlike last time. BECAUSE dear reader if... and IF I finish this before at least February/March I can actually start something for someone who could be crustal to the future of LOTTH! And this is dear reader is what is right now the deadline for myself! I want to keep the "LITERALLY AT THE END" section a bit back logged so it isn't gonna be rushed but everything else like main mechanics and logistic shall be done before the end of March! I do not want myself overworked, but at the same time if I let myself without a deadline I will stay with this logs for at least a year which could be good for you and for maximising the feature, but bad for other later plans and I need to learn how to work for a further deadline instead of a shorter and less hopeful deadline. I want this good just like you (and you are the only one who can decide if it actually worked), but I also need to learn this lesson sooner or later and as I stand now I have the ENG and the Schedule to hopefully out done myself on the holidays and finish a good part of the basic mechanics and I just only need to finish most of the polish and interaction of the other parts of the game and in the end around mid February I have a functional game that only needs debugging and polish and if everything goes right I have a fully polished and  useful mechanic out of the few that will be LOTTH's standing pillars!

In conclusion:
I have a lot of things to do I haven't even started with player choice and enemy AI yet but I already making promises, but I feel like I should be able to get it done and dusted by then! I feel positive and If all goes to shit I can always give myself a bit more deadline but for as a test it will be good enough!

Stay soon dear reader I hope the future and not blunder of the game Legend Of The True Heroes!