2024/12/08 18:38 Ability done and much follows
And.... DONE! Alright I finished ability with really junky mode, but for the time I test out other stuff it will work out. Next is Attack indicator, Choosing the player and... oh oh... TURN! OH SHIT

Alright alright I will not gonna over exaggerate it a bit, and it can be a problem IF I not keep it in check!

For now I did the following:
- Ability shows a card next to abilities
- Abilities gets pulled from a data canter and assigns automatically damage deals and health receives!
- Abilities texture gets pulled from folder
- Abilities have a handler but currently is kinda limited, expansion will be needed and it will 100% change in the future
- You can have a maximum of 18 Abilities with 2 different pages and if it less than 9 it will not turn to page 2 instead loop back around
- Cards have decks and they can be created really easily!
- Now I can focus on stuff easier!
- I can configure the range of the movement circle better!

and for right now I'm working on spawn points! I want to make it somewhat easy to use. I want 2 modes:
- Random
- Assigned to a named Character

random is explanatory, but Assigned is also kinda simple, the short answer is  that I want to add a name of a character to a spawn point and that character will gets to spawn exactly there. I'm also experimenting with more export variants so I can make my code a bit more interesting. Although it is needed to be noted that my code maximizing is not a 100% priority but a 65% at top! The reason is that I would love to make everything customizable and fun to edit, handy for future me, but for right atm the plan is to make sure that everything is at least have a 60-70% working rate that works how I want it at the moment, when I need to actually consider customize ability and other usefulness is something I worry about in a later date, for now I need stuff to work and If I can make it easier in the future in the moment I will try it, but I do not want to maximize or over complicate my project with useless what if's, bc they are not important at all.

PRO TIP!

if you make a game just make everything some what working so when in the future you actually need stuff again you will not gonna considering stabbing yourself with a first glens, but also you were able to proceed with your project! :)