2024/01/14 Attack Indicator base implumentation
So I've Implemented the Attack Indicator, it was kinda meh I still need to work on it like... I need to add crustal stuff to it to work like animations and actual ellipse calculations, but my problem with ellipse is still up and I still need some time before I feel like it that I can implement it correctly, and as of now it will stay like this...

I kinda feel like I have less and less time to do shittier and shittier implementations which isn't the best and I want to work on them, but I feel like I have other more important stuff and find really hard to sit down and just work on it with out giant problems rising in every step. I work slowly these couple of months now and it doesn't feel like I will finish until the end of March that much I still will try to do my best, but I feel like I should less talking and more working would solve my issue...

For now I will start to work on the actual first stuff I need to add that gonna make me tear out my hair... I will start working on turns... I have an idea what I want, but I feel like it is an even worse concept than the easier approach to one man one choice one chance and then the next guy, but I'm here to test and fiddle with ideas and I must say that I need to play around with this idea.

The short answer to what I'm thinking in case I don't actually make it is that:

I want to create an order to every choice you or the enemies make and then execute them at the end of everyone's round! The logic behind it is that the player could cancel at their last choice and go back and can change their mind/strategy even at the last second with ghost nodes that serve to get position, ability, action, bag and run data from the given orders and then hopefully use those to play out them at once with out too much hustle. The problems that I see already is already appearing right in front of my eyes and honestly the best idea would be if I would feature test the idea before I actually implement it like with the actual menu system. I see about it until the next commit.

I know that commits are more and more far between the reason is that I have to progressively add more and more stuff to it and making sure that the minimum of working together is okay is still a really hard task, and the fact that I have even more stuff than ever does not make it easier to deal with this projects.

In any case I still have time to finish this project although scares I will not stop and I really and I MEAN REALLY want to speed up soon. I will have to improvise a plan to make that happen but perhaps maybe if some stuff starts to settle down then I will be able to focus on this project more and more and in the end I will be able to make it into an actual reality BEFORE the end of March. Welp that's it for the update bye :D.