2024/1/27 21:58 Turns Halfway point! WOHOO
ALRIGHT! It's been... 2 WEEK????

Okay I've been a bit busy lately and recently when I actually started to feel the obligation to finish this project on the count that I will let all the other objectively more important things a bit on the hanging side, like Digital culture finals and this year ass whopping final with cisco and the stupid ass network administrator class (Which I feel like I have to say I have nothing against it I'm just way to shitty network administrator and I would not want to get ma money in the industry even if my life depends on it).

The point is that I been lacking a bit and slowly chipping away on this project on what time I have on the counter part of the others! BUT I've been working on it for the semi often majority of this 2 week. First with a prototype and then now with the actual project.

The prototype was a fucking success and now this is just the icing on the virtual cake!

It is still rough and a bunch of stuff like:
- actions
- amiable abilities
- movement

have not been implemented, but for the other like 3-4 stuff like target and self abilities and bag pack items it is a incredible that they are actually working!

How it works is that I have a node called the Turn node this has connections to a bunch of other nodes such as enemy, player groups, initiations, menu and a bunch more. And from there it has an _action() function that will go through a "stack" (aka the queued actions) and will execute the correct functions and sets the correct values for each options that the player choose in the order that the turn was set. This can dynamically change at anytime and we can even manipulate the enemies options and the order that everyone comes in!

In short it is a lot of fun and finally I can see the actual sauce of this project it is just FUN to press buttons and see eng go down or simple code I wrote get executed in order, and none of the less to say that it isn't an entire mess the code so I can still after 5 MONTHS I can still read the code fluidly and not get confused (too much) in my own code. It is a mess with shitty variables, confusing func names and other shitty if, else statements that doesn't make any sense and just idk shitty, but for now I can work with it and if I don't like a section I can happily rewrite it and still can hold the my ground against my own spaghetti code. It is a fucking miracle!

BUT for now as I stand I will head to bed and continue tomorrow ish time.

And for the deadline... well I - sigh ... you see the finals are closing in and today I realized that I have around 3 months before the school is over which is crazy bc I feel like I just started the school year like idk 2 month ago... this entire school year have been crazy fun and it is moving crazy fast but that is my main problem too before I realize it it has already been 5 MONTH since I started I already did 27 commits to this project and wrote around 30 scripts and thousands of lines of code and still we are far away from the actual funner part with animations, more attacks, polish and other END game stuff and for now it feels like it is an impossibility and it might be...

I don't want to give up, NO I will not give up until March and I DON'T CARE about anything, not about school, not about collage, not about my self... okay maybe a bit about my self but just as much so I don't look suicidal... THE POINT IS that I will not gonna let this project go and even if I cannot finish/polish it until March I will on the summer next to job, next on being ready for collage and next on the vacation in Greek I WILL do this if that is my last thing I have to endure... it is just that time moves quickly and I fear that I will not gonna have enough time before the deadline, and I have to resort to summer which at that point could be 2-4 month break and through that time I will forget a lot of my code and standards. So I should do everything I can before I have to put it down for school.

ALRIGHT, I wrote like more line than in this commit so I guess I should... stop it... okay I will go now I been writing this for like 30 min now LOL XD (it's 22:30 no oh shit XD)

Alright until next time ... :) <3