2025/2/07 09:54 Turn is done but a bit buggy :p
Alright I'm writing this from my school. I should do my assignment but instead I'm writing this commit bc I wanna work on the project on the mean time and Network management is not my style.

So I finished what left of the like movement and a "close" range abilities I'm still have to do the item removal and  the maybe the run option too, but for now they are in the back log and I just want to change scenery. So I will start on doing the enemy basics like:
- Attack
- Movement
- Actions
and then later down the line after the next commit I will do:
- Smart AI
- Death
- Spare

and from there I will need the ending of the battle by:
-Violence
or
- Mercy

and I thing almost that is it for this section, the hardest 100% goanna be the Enemy AI, bc I need to give editable logic to the enemy that can decide when to attack me, retreat, act randomly, use perhaps an item or just simply move to my direction. Giving all to an AI that cannot see or think on its own is gonna be a hustle and I feel like I just want to do the basic, which is are:
- Moving towards the player when...
- Attacking the player when...
- Retreating from the player when...
- Use Act if...
and much more if needed!

For now I would like to take stuff slower. The problem that come up this week is that I start to feel less an less attached and excited about this project. It feels like I'm no where near completion and the worst part is that this is only just a fraction of the whole game and its scope. I'm only doing the basic frame of the RPG mechanics, I want a whole movement mechanic on top of that and metroidvania like map  structure with branching paths and endings... and from day to day the sound of that feels more and more impossible alone. I WANT to do it and I will somehow get around it to make it. I just have to I want to, but the main problem of the right now is that I'm a solo dev trying to make a mechanic that feels ever so large to my skills, I felt this weeks beginning that I have no clue what I'm doing and my code is worthless, and I felt the option of giving up the first time in this entire life of me writing the story for the larger game. I had thoughts before hand but almost all of time they were like scaring thoughts that I CAN'T do it, but they went away after a couple of days or hours. This was a bit different seeing as my own code gets above me and almost push is me down and drowning me in my own ideas and solutions is gave me an actual stop to think:
"what if I actually cannot do it, what if all those people about gamedev is actually right and it is an impossibility for me a solo dev to create a game which is usually made by teams and studios of big companies?"
and that kinda shock me to the core...
I WANT to do it, but what if I CAN'T do it because it would take too much time and I would

A) I lose interest in the project after a while.
or
B) I give up because my code, my own brain starts to became a mess of nonsense and became unreadable, unmediatable jumble of mess.

I don't know what would I do if I have to give up my dream game, and I have to do it at the start of the journey...

I feel like only time could tell... and I can only hope that time will be merciful