2024/02/20 15:25 Some comments I written!
I commented some a few of the scripts, couldn't get to far.
For now I want to search for a fix for the radius of the ellipses!
I should have known this day will come and I should have been more prepared, but I feel like this is my biggest thorn in my side right now, I can figure out some band aid fix for the other bugs or some temporal ones, but this will not gonna just go away, I have to get to the end of it there is no free pass here... I have to start learning MATH and other ideas of tracking the mouse and when it is inside or close to something. Welp gotta go now see ya later... hopefully