2025/03/20 All the comments are done!
okay... I know you can see the dates between one commit and the other, and know if you even care at all whats up, but it's technically been a month since I changed everything (it was 15 but still I haven't changed a single real line of code since a month).

So what happend... not too much. I been trying to ignore the project and do study stuff cause I'm closing in to 2 month left of school and wanted to perhaps focus on it more, but... I got really tired of not finishing even the commenting and just felt real bad, so I just had to at least do this. I want to also do one more video on what I did so far so I can at least tell all the people I left in the dark since december how is the project going... I hope they still care at least a tiny bit.

as for me Its been a quiet month I definatly could have worked on the game this month with good faith cause as I said there wasn't even that much learning in this entire month, and I only kept saying to myself it is bc of burn out, we should learn and lets try numbing down the feeling for game making until january, but it was like pushing down an addiction, and trying to forget it and just make belive that you don't need it... I kinda needed it.

and with the fact I actually was burning out a bit and with the 1 1/2 week of sickness was just terrible!

Alright I have to go this 2 monkeys want me to leave so I have to cut it short... Idk when i will commit next, but I've been experimenting with a 3D alternative for the moveing cause the current system clearly does not work out.

(also I'm sorry for the shitty english I factory reseted my pc and I downloaded the HU version so every english line is wrong... whoops all wrong)