2025/06/16 15:41 changes to the structure and turn handler
BUT back on track, I actually did some stuff! First I changed the godot version from 4.3 to 4.4, tested some new turn handler logic, created an easier way to cross reference between scripts (it might be slower, like oh so much slower than any other way but I'm going for understanding and readability so idc), put the 3d version of LOTTH to the files so I might easier test it when I have the time and brain power. And finally I started to work on an order tester, because my handling of who's turn is kinda flawed! But that is about it, I'm still far away from adding any new stuff like enemy AI, but for now this is progress too, so I'm not sad that I'm going in a snails neck pace, but I rather looking at this as fun after work (when I have the brain capacity ofc)!