2025/06/20-21 12:11 Big change of plan and some testing
I'm writing this at midnight and I have no clue which date should I choose, but oh well, why not both!

Anyways I've done some testing, and came to the conclusion that I need (WAIT I JUST FOUND WHERE TO EABLE THE SPELL CHECK OMG) to rework how my characters interact one after the other, mainly because there are some bugs that are months old at this point and could be hard for me to snipe them down one by one, and in the meanwhile found a sort of better way to do it myself... not sure have to test it out, but for now that is my main priority. AND the big change IS...

drum roll please...

I will put moving into the back burner... 🎊 🎊🎊🎊

For further clarification, I cannot do the math behind movement at the moment because it is not a circle but an ellipsis mainly because of the isometric view, and right now it would take me way too long to start learning everything about ellipses, and make the movement and aiming actually possible, it is but later. For now I want to focus as I said, on the turns and other small bugs, making a win/lose condition and focusing on enemy AI. I want to make a game after all and if I side track about one feature than all the others, once again would get neglected, which would result in more set back and reworks, SO until I can start learning ellipsis theory, I will focus on the actual meats and potatoes of the game and around late July early Augustus I will start learning both C and Math. That's right your boy will start learning C as a MAN or BIG BOY. We will C ;D

Anyways I will go to bed now, for now I will just start the testing of the new idea I had, and slowly building it into the core mechanics. I should also rebuild the god damn Trello sheet, but it is on THE LIST :trollface:

ANYWAYS I will go and think about all that stuff.
See you next update!